Adds You Tube menu
Swisher cards can be created from the right hand menu.

If the user gives permission their recent Likes & Favorites are
shown and these can be associated with a swisher card.

The playing of youtube videos is not part of this commit.

To get access to the user's Likes & Favorites the user must:
 -update play-services
 -give permission
If these steps have not been completed the menu shows an
action which will take the through the process.

For swisher to use the youtube v3 api there must be a registered
google application which has the youtube v3 api enabled and an
android oauth2 credential with a package name matching the
applicationId package name in build.gradle and an sha matching
the sha of the keystore used to build the apk